Revert "separated UI from backboard configuration"
This reverts commit 4124c3ae24a976c86fa9e8aee4f4656cb08b5ef2.
