+++ bundles/de.uka.ipd.sdq.workflow.mdsd/src/de/uka/ipd/sdq/workflow/mdsd/AbstractWorkflowBasedMDSDLaunchConfigurationDelegate.java
@@ -1,10 +1,16 @@
 package de.uka.ipd.sdq.workflow.mdsd;
 
+import java.util.List;
+
+import org.apache.log4j.Level;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.debug.core.ILaunch;
 
+import de.uka.ipd.sdq.workflow.BlackboardBasedWorkflow;
+import de.uka.ipd.sdq.workflow.launchconfig.core.AbstractWorkflowBasedLaunchConfigurationDelegate;
 import de.uka.ipd.sdq.workflow.launchconfig.core.AbstractWorkflowBasedRunConfiguration;
+import de.uka.ipd.sdq.workflow.logging.console.LoggerAppenderStruct;
 import de.uka.ipd.sdq.workflow.mdsd.blackboard.MDSDBlackboard;
 import de.uka.ipd.sdq.workflow.mdsd.blackboard.ResourceSetPartition;
 import de.uka.ipd.sdq.workflow.ui.UIBasedWorkflow;
@@ -20,7 +26,17 @@ import de.uka.ipd.sdq.workflow.ui.UIBasedWorkflow;
  * @author Steffen Becker
  */
 public abstract class AbstractWorkflowBasedMDSDLaunchConfigurationDelegate<WorkflowConfigurationType extends AbstractWorkflowBasedRunConfiguration>
-        extends AbstractWorkflowBasedMDSDBlackboardLaunchConfigurationDelegate<WorkflowConfigurationType> {
+        extends
+        AbstractWorkflowBasedLaunchConfigurationDelegate<WorkflowConfigurationType, BlackboardBasedWorkflow<MDSDBlackboard>> {
+
+    /**
+     * Factory method for the blackboard used in the workflow of this launch.
+     *
+     * @return The blackboard to be used in the workflow
+     */
+    protected MDSDBlackboard createBlackboard() {
+        return new MDSDBlackboard();
+    }
 
     /*
      * (non-Javadoc)
@@ -36,4 +52,19 @@ public abstract class AbstractWorkflowBasedMDSDLaunchConfigurationDelegate<Workf
                 this.createExceptionHandler(workflowConfiguration.isInteractive()), this.createBlackboard());
     }
 
+    /*
+     * (non-Javadoc)
+     *
+     * @see de.uka.ipd.sdq.workflow.launchconfig.AbstractWorkflowBasedLaunchConfigurationDelegate#
+     * setupLogging(org.apache.log4j.Level)
+     */
+    @Override
+    protected List<LoggerAppenderStruct> setupLogging(final Level logLevel) throws CoreException {
+        List<LoggerAppenderStruct> loggerList = super.setupLogging(logLevel);
+
+        // Setup openArchitectureWare Logging
+        loggerList.add(this.setupLogger("org.openarchitectureware", logLevel, SHORT_LOG_PATTERN));
+
+        return loggerList;
+    }
 }
